Both test passing
