+++ drools-model/drools-model-compiler/src/main/java/org/drools/modelcompiler/builder/generator/visitor/accumulate/AccumulateVisitorPatternDSL.java
@@ -17,6 +17,7 @@ import org.drools.modelcompiler.builder.generator.DrlxParseUtil;
 import org.drools.modelcompiler.builder.generator.RuleContext;
 import org.drools.modelcompiler.builder.generator.expression.PatternExpressionBuilder;
 import org.drools.modelcompiler.builder.generator.visitor.ModelGeneratorVisitor;
+import org.drools.modelcompiler.util.LambdaUtil;
 
 import static org.drools.modelcompiler.builder.generator.DrlxParseUtil.findPatternWithBinding;
 import static org.drools.modelcompiler.builder.generator.DrlxParseUtil.fromVar;
@@ -58,6 +59,23 @@ public class AccumulateVisitorPatternDSL extends AccumulateVisitor {
             }
         });
         return Optional.empty();
+    }
+
+    private void composeTwoBindings(MethodCallExpr newBindingExpression, MethodCallExpr pattern) {
+        pattern.getParentNode().ifPresent(oldBindExpression -> {
+            MethodCallExpr oldBind = (MethodCallExpr) oldBindExpression;
+
+            LambdaExpr oldBindLambda = (LambdaExpr) oldBind.getArgument(1);
+            LambdaExpr newBindLambda = (LambdaExpr) newBindingExpression.getArgument(1);
+
+            Expression newComposedLambda = LambdaUtil.compose(oldBindLambda, newBindLambda);
+
+            newBindingExpression.getArguments().removeLast();
+            newBindingExpression.addArgument(newComposedLambda);
+
+            newBindingExpression.setScope(pattern);
+            oldBind.replace(newBindingExpression);
+        });
     }
 
     private void addBindAsLastChainCall(MethodCallExpr newBindingExpression, MethodCallExpr pattern) {
